requirements.txt
streamlit
sentence-transformers
transformers
torch
yt-dlp
openai-whisper
numpy
scikit-learn
nltk
